% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_tools.R
\name{ev_disconnect}
\alias{ev_disconnect}
\title{Disconnect from EpiVault}
\usage{
ev_disconnect(con)
}
\arguments{
\item{con}{An EpiVault connection object}
}
\description{
Disconnects from EpiVault instance identified by \code{con}
}
\examples{
\dontrun{

con <- ev_connect()

ev_disconnect(con)

}
}
\seealso{
Other {EpiVault connectors}: 
\code{\link{ev_connect}()}
}
\concept{{EpiVault connectors}}
