% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{ev_simple_fetch}
\alias{ev_simple_fetch}
\title{Simple data fetcher}
\usage{
ev_simple_fetch(con, project, table, visibility = 0, variables = character(0))
}
\arguments{
\item{con}{An EpiVault connection object}

\item{project}{Project name}

\item{table}{Table name}

\item{visibility}{(default=0) Maximum visibility level requested. See examples.}

\item{variables}{(optional) Character vector of variable names. These should just be the variable (column) names, i.e. not fully qualified. If omitted, all variables are returned.}
}
\value{
A data frame containing the data requested
}
\description{
Fetches data from a single table in EpiVault and returns it as a data frame.
}
\examples{
\dontrun{

con <- ev_connect()

# fetch the whole table
dat <- ev_simple_fetch(con,
                       project = "project1",
                       table = "table1",
                       visibility = 0
)

# just fetch a couple of variables
# NB. variables in the table tagged as 'required' will also be returned
dat <- ev_simple_fetch(con,
                       project = "project1",
                       table = "table1",
                       visibility = 0,
                       variables = c("var11", "var13", "var44")
)

# use higher visibility level to access sensitive variables
# special database permissions may be required 
dat <- ev_simple_fetch(con,
                       project = "project1",
                       table = "table1",
                       visibility = 9,
                       variables = c("date_of_birth")
)
                       
}

}
\seealso{
Other {data request functions}: 
\code{\link{fetch_ev_data}()},
\code{\link{fetch_ev_meta_tabs}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{make_ev_variables}()},
\code{\link{read_ev_variables}()},
\code{\link{write_ev_data}()}
}
\concept{{data request functions}}
