% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{fetch_ev_data}
\alias{fetch_ev_data}
\title{Fetch data request from EpiVault}
\usage{
fetch_ev_data(con, ev_vars, visibility = 0)
}
\arguments{
\item{con}{An EpiVault connection object}

\item{ev_vars}{An \code{ev_variables} container}

\item{visibility}{(default=0) Maximum visibility level requested. See examples.}
}
\value{
An \code{ev_data} container
}
\description{
Interprets an \code{ev_variables} container, fetches the data from EpiVault and returns it in an \code{ev_data} container.
}
\details{
Passes \code{ev_vars} to \code{\link{fetch_ev_meta_vars}} and \code{\link{fetch_ev_meta_tabs}} to get the metadata ready for processing, fetches the tabular data from EpiVault before applying data labels. All fetched data and metadata is returned in an \code{ev_data} container.
}
\examples{
\dontrun{

con <- ev_connect()

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")

# fetch the data from EpiVault - default visibility for standard data requests
ev <- fetch_ev_data(con, vars)

# fetch the data from EpiVault - highest visibility to access sensitive variables
ev <- fetch_ev_data(con, vars, visibility = 9)

}

}
\seealso{
Other {data request functions}: 
\code{\link{ev_simple_fetch}()},
\code{\link{fetch_ev_meta_tabs}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{make_ev_variables}()},
\code{\link{read_ev_variables}()},
\code{\link{write_ev_data}()}
}
\concept{{data request functions}}
