% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{fetch_ev_meta_tabs}
\alias{fetch_ev_meta_tabs}
\title{Fetch table metadata from EpiVault}
\usage{
fetch_ev_meta_tabs(con, ev_vars)
}
\arguments{
\item{con}{An EpiVault connection object}

\item{ev_vars}{An \code{ev_variables} container}
}
\value{
A data frame
}
\description{
Interprets an \code{ev_variables} container, fetches the table metadata from EpiVault and returns it in a data frame.
}
\examples{
\dontrun{

con <- ev_connect()

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")

# fetch table metadata from EpiVault
meta <- fetch_ev_meta_tabs(con, vars)

}

}
\seealso{
Other {data request functions}: 
\code{\link{ev_simple_fetch}()},
\code{\link{fetch_ev_data}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{make_ev_variables}()},
\code{\link{read_ev_variables}()},
\code{\link{write_ev_data}()}
}
\concept{{data request functions}}
