% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_data}
\alias{get_ev_data}
\title{Get single data frame}
\usage{
get_ev_data(ev_data, df_name = character(0), df_index = integer(0))
}
\arguments{
\item{ev_data}{An \code{ev_data} container}

\item{df_name}{The name of the data frame. Only used if \code{df_index} is not supplied}

\item{df_index}{The numerical index of the data frame within the \code{ev_data} container. Overrides \code{df_name} if both are supplied.}
}
\value{
A data frame
}
\description{
Returns a selected data frame from an \code{ev_data} container.
}
\details{
The data frame can be selected by index (1, 2, 3 etc.) or by name (\code{project.table}). The names of data frames within an \code{ev_data} container can be queried using \code{\link{get_ev_data_names}}.

An \code{ev_data} container can be created using \code{\link{fetch_ev_data}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path-to-variables-file")
# fetch the data from EpiVault
dat <- fetch_ev_data(vars)

# get a data frame by name
get_ev_data(dat, df_name = "proj1.tab1")
# get a data frame by index
get_ev_data(dat, df_index = 1)

}

}
\seealso{
\code{\link{fetch_ev_data}} to create an \code{ev_data} container.

Other {getters}: 
\code{\link{get_ev_data_names}()},
\code{\link{get_ev_metadata}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_variables}()},
\code{\link{get_ev_vars_df}()}
}
\concept{{getters}}
