% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_data_names}
\alias{get_ev_data_names}
\title{Get table/data frame names}
\usage{
get_ev_data_names(ev_data)
}
\arguments{
\item{ev_data}{An \code{ev_data} container}
}
\value{
A character vector of data names
}
\description{
Returns the names of the data frames in an \code{ev_data} container.
}
\details{
The returned values can be used in a call to \code{\link{get_ev_data}}. 

An \code{ev_data} container can be created using \code{\link{fetch_ev_data}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path-to-variables-file")
# fetch the data from EpiVault
dat <- fetch_ev_data(vars)

# get the data frame names
get_ev_data_names(dat)

}

}
\seealso{
\code{\link{fetch_ev_data}} to create an \code{ev_data} container.

Other {getters}: 
\code{\link{get_ev_data}()},
\code{\link{get_ev_metadata}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_variables}()},
\code{\link{get_ev_vars_df}()}
}
\concept{{getters}}
