% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_metadata}
\alias{get_ev_metadata}
\title{Get single metadata table}
\usage{
get_ev_metadata(ev_data, type = character(0))
}
\arguments{
\item{ev_data}{An \code{ev_data} container}

\item{type}{One of \code{c('variable', 'category', 'table')}. If not recognised, returns an empty data frame.}
}
\value{
A data frame of metadata
}
\description{
Returns metadata of a selected type from an \code{ev_data} container.
}
\details{
Three types of metadata can be returned: 

\describe{
  \item{\emph{variable}}{Variable level information such as name, label, description, keywords and summary stats.}
  \item{\emph{category}}{Value labels for categorical variabbles, one row per value. Returns variable name, value and label.}
  \item{\emph{table}}{Table level information such as name, cohort membership, entity type, subjects and respondents and table description.}
}

An \code{ev_data} container can be created using \code{\link{fetch_ev_data}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path-to-variables-file")
# fetch the data from EpiVault
dat <- fetch_ev_data(vars)

# get variable metadata
get_ev_metadata(dat, type = "variable")
# get category metadata
get_ev_metadata(dat, type = "category")
# get table metadata
get_ev_metadata(dat, type = "table")

}

}
\seealso{
\code{\link{fetch_ev_data}} to create an \code{ev_data} container.

Other {getters}: 
\code{\link{get_ev_data_names}()},
\code{\link{get_ev_data}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_variables}()},
\code{\link{get_ev_vars_df}()}
}
\concept{{getters}}
