% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_projects}
\alias{get_ev_projects}
\title{Get project names}
\usage{
get_ev_projects(ev_vars)
}
\arguments{
\item{ev_vars}{An `ev_variables` container}
}
\value{
A character vector of project names
}
\description{
Returns project names from an `ev_variables` container. An `ev_variables` container can be created using `read_ev_variables`.
}
\examples{
\dontrun{

vars <- read_ev_variables("path-to-variables-file")

get_ev_projects(vars)

}

}
