% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_request}
\alias{get_ev_request}
\title{Get ev_variables request}
\usage{
get_ev_request(ev_data)
}
\arguments{
\item{ev_data}{An \code{ev_data} container}
}
\value{
An \code{ev_variables} container
}
\description{
Returns the \code{ev_variables} container used to create an \code{ev_data} container.
}
\details{
Every \code{ev_data} container stores a copy of the \code{ev_variables} container used to create it, and this function is used to access it directly.
}
\examples{
\dontrun{

con <- ev_connect()

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")
# fetch the data from EpiVault
dat <- fetch_ev_data(con, vars)

# get variable request - a copy of the ev_variables container (vars) used to build dat
get_ev_request(dat)

}

}
\seealso{
\code{\link{fetch_ev_data}} to create an \code{ev_data} container.

Other {getters}: 
\code{\link{get_ev_data_names}()},
\code{\link{get_ev_data}()},
\code{\link{get_ev_metadata}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_variables}()},
\code{\link{get_ev_vars_df}()}
}
\concept{{getters}}
