% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_variables}
\alias{get_ev_variables}
\title{Get variables names}
\usage{
get_ev_variables(ev_vars)
}
\arguments{
\item{ev_vars}{An \code{ev_variables} container}
}
\value{
A character vector of variable names
}
\description{
Returns fully qualified variable names from an \code{ev_variables} container.
}
\details{
An \code{ev_variables} container can be created using \code{\link{read_ev_variables}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path-to-variables-file")

# get the variable names from the request
get_ev_variables(vars)

}

}
\seealso{
\code{\link{read_ev_variables}} to create an \code{ev_variables} container.

Other {getters}: 
\code{\link{get_ev_data_names}()},
\code{\link{get_ev_data}()},
\code{\link{get_ev_metadata}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_vars_df}()}
}
\concept{{getters}}
