% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_getters.R
\name{get_ev_vars_df}
\alias{get_ev_vars_df}
\title{Get variables data frame}
\usage{
get_ev_vars_df(ev_vars)
}
\arguments{
\item{ev_vars}{An \code{ev_variables} container}
}
\value{
A data frame containing fully qualified variable name, project name, table name and variable name.
}
\description{
Returns the variable data frame from an \code{ev_variables} container.
}
\details{
The variable data frame contains four columns:

\describe{
  \item{\emph{varfullname}}{Fully qualified variable name, i.e. \code{project.table.variable}}
  \item{\emph{project}}{Project name}
  \item{\emph{table}}{Table name}
  \item{\emph{variable}}{Variable name}
}

An \code{ev_variables} container can be created using \code{\link{read_ev_variables}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")

# get the variable data frame from the request
get_ev_vars_df(vars)

}

}
\seealso{
\code{\link{read_ev_variables}} to create an \code{ev_variables} container.

Other {getters}: 
\code{\link{get_ev_data_names}()},
\code{\link{get_ev_data}()},
\code{\link{get_ev_metadata}()},
\code{\link{get_ev_projects}()},
\code{\link{get_ev_tables}()},
\code{\link{get_ev_variables}()}
}
\concept{{getters}}
