% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{make_ev_variables}
\alias{make_ev_variables}
\title{Make ev_variables container}
\usage{
make_ev_variables(vars)
}
\arguments{
\item{vars}{Character vector of fully qualified variable names}
}
\value{
An \code{ev_variables} container
}
\description{
Converts a character vector of fully qualified variable names into an \code{ev_variables} container.
}
\details{
This is called by data request functions such as \code{\link{read_ev_variables}} and \code{\link{ev_simple_fetch}} and you would not normally call it directly. However, there may be circumstances in which all you have is a list of variable names and you wish to convert these into an \code{ev_variables} container.
}
\examples{
\dontrun{

vars <- c("proj1.tab1.var1", "proj1.tab1.var2", "proj2.tab2.var1")
make_ev_variables(vars)

}

}
\seealso{
Other {data request functions}: 
\code{\link{ev_simple_fetch}()},
\code{\link{fetch_ev_data}()},
\code{\link{fetch_ev_meta_tabs}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{read_ev_variables}()},
\code{\link{write_ev_data}()}
}
\concept{{data request functions}}
