% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{read_ev_variables}
\alias{read_ev_variables}
\title{Read variables from a file}
\usage{
read_ev_variables(file)
}
\arguments{
\item{file}{Path to the file to be read}
}
\value{
An \code{ev_variables} container
}
\description{
Reads variables from a file into an \code{ev_variables} container. Supports delimited or fixed width text and Excel formats. Will try to guess format based on file extension and first few lines.
}
\details{
Builds a character vector of fully qualified variable names and passes this to \code{\link{make_ev_variables}}.
}
\examples{
\dontrun{

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")

}

}
\seealso{
If you want to create an \code{ev_variables} container directly from a vector of variable names, see \code{\link{make_ev_variables}}.

Other {data request functions}: 
\code{\link{ev_simple_fetch}()},
\code{\link{fetch_ev_data}()},
\code{\link{fetch_ev_meta_tabs}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{make_ev_variables}()},
\code{\link{write_ev_data}()}
}
\concept{{data request functions}}
