% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_processing.R
\name{write_ev_data}
\alias{write_ev_data}
\title{Write data files}
\usage{
write_ev_data(
  ev_data,
  path,
  name = "ev_data",
  format = "stata",
  metadata = FALSE
)
}
\arguments{
\item{ev_data}{An \code{ev_data} container}

\item{path}{Path to the folder to save files to}

\item{name}{(default='ev_data') A name for the data request. Used as the prefix for every output file.}

\item{format}{(default='stata') File format to write to. One of \code{c('stata', 'csv')}.}

\item{metadata}{(default=FALSE) If \code{TRUE}, also output files for variable, category and table metadata.}
}
\description{
Writes the data and metadata from an \code{ev_data} container to the specified file format.
}
\examples{
\dontrun{

con <- ev_connect()

# read a variable data request file
vars <- read_ev_variables("path/to/variables/file")

# fetch the data from EpiVault
ev <- fetch_ev_data(con, vars)

# write the data files in default stata format, excluding metadata
write_ev_data(ev_data = ev,
              path = "path/to/data/request/folder",
              name = "request101"
)

# write the data files in csv format, including metadata
write_ev_data(ev_data = ev,
              path = "path/to/data/request/folder",
              name = "request101",
              format = 'csv',
              metadata = TRUE
)

}

}
\seealso{
Other {data request functions}: 
\code{\link{ev_simple_fetch}()},
\code{\link{fetch_ev_data}()},
\code{\link{fetch_ev_meta_tabs}()},
\code{\link{fetch_ev_meta_vars}()},
\code{\link{make_ev_variables}()},
\code{\link{read_ev_variables}()}
}
\concept{{data request functions}}
