% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epivaultr_tools.R
\name{ev_disconnect}
\alias{ev_disconnect}
\title{Disconnect from EpiVault}
\usage{
ev_disconnect(con)
}
\arguments{
\item{con}{EpiVault connection}
}
\description{
Disconnects from EpiVault instance identified by `con`
}
\examples{
\dontrun{

con <- ev_connect()

ev_disconnect(con)

}
}
